:recycle: refactor: 관련 있는 로직끼리 패키지(폴더)로 묶기 - player(Computer/User) - game(Game/ErrorMsg/Print/Validate)
