+++ src/main/java/baseball/game/Validate.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.game;
 
 import java.util.List;
 import java.util.regex.Pattern;
