+++ src/main/java/baseball/player/Computer.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.player;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
